+++ common/src/main/java/jimenezli/neuro21/util/EntityNames.java
@@ -2,6 +2,8 @@ package jimenezli.neuro21.util;
 
 public class EntityNames {
     public static final String VEDAL = "vedal";
+    public static final String GIGA_VEDAL = "giga_vedal";
+    public static final String MIYUNE = "miyune";
     public static final String ANNY = "anny";
     public static final String NEUROSAMA = "neurosama";
     public static final String EVIL_NEUROSAMA = "evil_neurosama";